% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcQualMeasures.R
\name{calcMatrixScore}
\alias{calcMatrixScore}
\title{This function calculates score of a matrix as a sum of partial scores of each columns}
\usage{
calcMatrixScore(matrix, covered.pos)
}
\arguments{
\item{matrix}{object to calculate score for}

\item{covered.pos}{vector of snv positions to consider}
}
\description{
This function calculates score of a matrix as a sum of partial scores of each columns
}
\author{
David Porubsky
}

