% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/callBreakpoint.R
\name{callBreakpointR}
\alias{callBreakpointR}
\title{This function will call BreakPointR on phased reads object}
\usage{
callBreakpointR(datapath, outputfolder = "BreakPointR")
}
\arguments{
\item{datapath}{}

\item{outputfolder}{}
}
\description{
This function will call BreakPointR on phased reads object
}
\author{
David Porubsky
}

