% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareSingleCellHaps.R
\name{compareSingleCellHaps}
\alias{compareSingleCellHaps}
\title{This will take sorted matrices and will calculate concordance of each single cell to the consensus haplotypes in order to assemble highly 
accurate haplotypes}
\usage{
compareSingleCellHaps(consensusHaps = NULL, sortedHaps = NULL,
  bin.size = 10, step = 1)
}
\arguments{
\item{consensusHaps}{A \code{list} returned by function \code{\link{assembleHaps}}}

\item{sortedHaps}{A \code{list} returned by function \code{\link{sortMatrices}}}

\item{bin.size}{A size of the window measured in number of SNV to scan single cell haplotypes.}

\item{step}{A step (number of SNVs) to move the window.}
}
\description{
This will take sorted matrices and will calculate concordance of each single cell to the consensus haplotypes in order to assemble highly 
accurate haplotypes
}
\author{
David Porubsky
}

