% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctInversionPhasing.R
\name{correctHomInv}
\alias{correctHomInv}
\title{Correct homozygous inversion in StrandPhaseR VCF file.}
\usage{
correctHomInv(correct.gr = NULL, vcf.file = NULL, ID = "")
}
\arguments{
\item{correct.gr}{A \code{\link{GRanges}} object of homozygous inversion site to be corrected.}

\item{vcf.file}{A \pkg{\link{StrandPhaseR}} formatted VCF file to be corrected for inversion phasing.}

\item{ID}{A unique id to be appended at the end of each corrected VCF file.}
}
\value{
\code{NULL}
}
\description{
This function takes as an input region deemed to be a homozygous inversion and flips H1 and H2 haplotypes for SNVs that 
overlap with this region.
}
\author{
David Porubsky
}
