% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctInversionPhasing.R
\name{correctInvertedRegionPhasing}
\alias{correctInvertedRegionPhasing}
\title{Correct StrandPhaseR phasing over inverted regions}
\usage{
correctInvertedRegionPhasing(
  input.bams,
  outputfolder = NULL,
  inv.bed = NULL,
  recall.phased = FALSE,
  het.genotype = "strict",
  chromosomes = NULL,
  snv.positions = NULL,
  composite.file = NULL,
  breakpointR.data = NULL,
  strandphaseR.data = NULL,
  pairedEndReads = TRUE,
  min.mapq = 10,
  vcfs.files = NULL,
  lookup.bp = 1e+06,
  lookup.blacklist = NULL,
  bsGenome = NULL,
  ref.fasta = NULL,
  assume.biallelic = TRUE,
  overwrite.results = TRUE
)
}
\arguments{
\item{input.bams}{A path to a VCF file to be loaded.}

\item{outputfolder}{A path to a folder where all summary and intermediate results will be stored.}

\item{inv.bed}{A BED formatted file that contains likely inverted regions to evaluate.}

\item{recall.phased}{If set to \code{TRUE},inversion signatures will be re-called using phased Strand-seq reads (Useful to redefine HET inversion boundaries).}

\item{het.genotype}{If set to 'strict' heterozygous inversion has to be supported by both 'breakpointR.data' and 'strandphaseR.data',
if set to 'lenient' support from either 'breakpointR.data' only is sufficient.}

\item{chromosomes}{Limit analysis to a certain chromosomes only.}

\item{snv.positions}{A path to a VCF files containing SNV positions to phase.}

\item{composite.file}{...}

\item{breakpointR.data}{A path to results obtained by running \code{\link{breakpointR}} package on a given sample.}

\item{strandphaseR.data}{A path to results obtained by running \code{\link{StrandPhaseR}} package on a given sample.}

\item{pairedEndReads}{Set to \code{TRUE} if you have paired-end reads in your file.}

\item{min.mapq}{Minimum mapping quality when importing from BAM files.}

\item{vcfs.files}{A path to a folder where all VCF files to be corrected are stored (sample specific).}

\item{lookup.bp}{A number of nucleotides, downstream and upstream, from the heterozygous inversion site ('phase.gr') to be genotyped.}

\item{lookup.blacklist}{A \code{\link{GRanges}} object or a path to a BED file containing a set of ranges to be excluded 
when extending 'phase.gr' by 'lookup.bp'. The total size of 'lookup.bp' is kept after filtering.}

\item{assume.biallelic}{If set to \code{TRUE} parameter 'snv.positions' is expected to contain biallelic loci (0/1, 1/0) and thus
gaps in haplotypes will be filled accordingly.}

\item{overwrite.results}{Set to \code{FALSE} if existing corrected VCF files should be used instead of creating a new copy of original VCF file.}
}
\value{
\code{NULL} or \code{data.frame} depending if 'outputfolder' is defined.
}
\description{
This function attempts to correct phasing over inverted regions using both \pkg{\link{breakpointR}} and \pkg{\link{StrandPhaseR}} results.
}
\author{
David Porubsky
}
