% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportConsensus.R
\name{exportConsensus}
\alias{exportConsensus}
\title{This funcion will load sorted watson and crick matrices and will collapsed them in order to obtain consensus string of each matrix}
\usage{
exportConsensus(data.bases, data.quals, min.cov = 2, translateBases = FALSE)
}
\arguments{
\item{min.cov}{dtermines minimal number of cells covering each position to consider in consensus}

\item{translateBases}{translates integer coded bases (1,2,3,4) into letters (A,C,G,T)}

\item{data.object}{containing sorted watson and crick haplotypes of each single cell}
}
\description{
This funcion will load sorted watson and crick matrices and will collapsed them in order to obtain consensus string of each matrix
}
\author{
David Porubsky
}

