% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportPhasedReads.R
\name{exportPhasedReads}
\alias{exportPhasedReads}
\title{Export Strand-seq phased reads in FASTA or FASTQ format.}
\usage{
exportPhasedReads(
  phased.tab = NULL,
  inputfolder = NULL,
  outputfolder = NULL,
  by.matepair = TRUE,
  min.mapq = 10,
  pairedEndReads = TRUE,
  format = "fastq"
)
}
\arguments{
\item{phased.tab}{A data table containing haplotype assignment for each single cell and haplotype informative region.
This table is standard output of \pkg{StrandPhaseR} and cen be found in subfolder 'Phased/phased_haps.txt'.}

\item{inputfolder}{Path to the bam files to process}

\item{outputfolder}{Output directory. If non-existent it will be created.}

\item{by.matepair}{If set to \code{TRUE} (the default) paired-end reads will be exported per mate-pair (read1 and read2).}

\item{min.mapq}{Minimum mapping quality when importing from BAM files.}

\item{pairedEndReads}{Set to \code{TRUE} if you have paired-end reads in your file.}

\item{format}{User defined output format, either "fasta" or "fastq" (the default).}
}
\description{
This function will take phase information for each single cell and each haplotype informative region 
and will divide all mapped reads in BAM format into haplotype specific sets. In case of paired-end 
reads, read1 and read2 will be exported in separate files.
}
\author{
David Porubsky
}
