% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillGaps.R
\name{fillGapsWithBam}
\alias{fillGapsWithBam}
\title{This function will take both assembled haplotypes and will try to fill gaps at positions where only one allele is phased
Such position have to be heterozygous so the alternative allele at this position can be reliably distinguished}
\usage{
fillGapsWithBam(data.object, merged.bam, min.mapq = 10, min.baseq = 30,
  translateBases = FALSE, chromosome = NULL, chunkSize = 1e+07,
  filterAltAlign = TRUE)
}
\arguments{
\item{data.object}{...}

\item{merged.bam}{...}

\item{min.mapq}{Minimum mapping quality when importing from BAM files.}

\item{min.baseq}{Minimum base quality to consider a base for phasing.}

\item{translateBases}{translates integer coded bases (1,2,3,4) into letters (A,C,G,T)}

\item{chromosome}{If only a subset of the chromosomes should be processed, specify them here.}

\item{chunkSize}{...}
}
\description{
This function will take both assembled haplotypes and will try to fill gaps at positions where only one allele is phased
Such position have to be heterozygous so the alternative allele at this position can be reliably distinguished
}
\author{
David Porubsky
}
