% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapRecomb.R
\name{findRecomb}
\alias{findRecomb}
\title{Find meiotic recombination breakpoints using circular binary segmentation (CBS)}
\usage{
findRecomb(comparison = NULL, minSeg = 100, smooth = 3, collapse.amb = TRUE)
}
\arguments{
\item{comparison}{A \code{\link{VRanges-class}} object}

\item{minSeg}{Minimal length (number of variants) being reported as haplotype block (\code{fastseg} parameter).}

\item{smooth}{Number of consecutive variants being considered as a random error and so being corrected (flipped).}

\item{collapse.amb}{Set to \code{TRUE} if segments with ambiguous haplotype assignments should be collapsed.}
}
\description{
This function takes as an input a \code{\link{VRanges-class}} object with extra metacolumn that contains
for each variant parantal identity. P1 -> parent1 & P2 -> parent2.
}
\author{
David Porubsky
}
