% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypeStrandScells.R
\name{genotypeStrandScells}
\alias{genotypeStrandScells}
\title{Genotype Strand-seq cell based on population genotypes.}
\usage{
genotypeStrandScells(
  inputfolder = NULL,
  strandS.vcf = NULL,
  popul.vcf.list = NULL,
  wc.regions = NULL,
  chromosomes = paste0("chr", c(1:22)),
  min.snv.cov = 5,
  max.snv.cov = 30,
  max.snv.per.chr = 30000,
  blacklist = NULL
)
}
\arguments{
\item{inputfolder}{A data folder where individual Strand-seq BAM files are stored.}

\item{strandS.vcf}{A VCF file that contains SNVs called from merged Strand-seq libraries using RTG tool.}

\item{popul.vcf.list}{A names list of paths to VCF files (per chromosome) that contains SNVs from multiple individuals such as 1000G sample panel.}

\item{wc.regions}{A Watson-crick regions per library and per chromosome as defined by breakpointR.}

\item{chromosomes}{List of chromosomes to be used for genotyping.}

\item{min.snv.cov}{A minumum number of Strand-seq reads required to cover a SNV position in 'strandS.vcf'.}

\item{max.snv.cov}{A maximum number of Strand-seq reads allowed to cover a SNV position in 'strandS.vcf'.}

\item{max.snv.per.chr}{A maximum number of SNVs to be loaded from population panel VCF ('popul.vcf) per chromosome.}

\item{blacklist}{A\code{\link{GRanges-class}} object of genomic regions to filter out SNV positions.}
}
\value{
A \code{data.frame} object.
}
\description{
This function will import variants (SNVs) covered in individual Strand-seq libraries in a form of BAM files aligned to a
a reference genome such as GRCh38. Such SNVs are compared to a population genotypes such as 1000G sample panel.
}
\author{
David Porubsky
}
