% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{loadGRangesFromFiles}
\alias{loadGRangesFromFiles}
\title{Load phased data from RData files}
\usage{
loadGRangesFromFiles(files)
}
\arguments{
\item{files}{A list of files that contain phased data.}
}
\value{
A list() containing all loaded phased data.
}
\description{
Load phased data from RData files
}
\author{
David Porubsky
}

