% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOHseeker.R
\name{LOHseeker}
\alias{LOHseeker}
\title{This function uses circular-binary-segementation to segment single cell haplotypes based on their similarity to consensus haploypes of all cells.}
\usage{
LOHseeker(data.object = NULL, chromosome = NULL, bin.size = 10)
}
\arguments{
\item{data.object}{A \code{list} output from the function \code{\link{compareSingleCellHaps}}}

\item{chromosome}{A chromosome id.}
}
\description{
This function uses circular-binary-segementation to segment single cell haplotypes based on their similarity to consensus haploypes of all cells.
}
\author{
David Porubsky
}
