% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseChromsome.R
\name{phaseChromosome}
\alias{phaseChromosome}
\title{Wrapper function for StrandPhaseR to phase a single chromosome.}
\usage{
phaseChromosome(
  inputfolder,
  outputfolder = "./StrandPhaseR_analysis",
  positions = NULL,
  WCregions = NULL,
  chromosome = NULL,
  pairedEndReads = TRUE,
  min.mapq = 10,
  min.baseq = 30,
  num.iterations = 2,
  translateBases = TRUE,
  concordance = 0.9,
  fillMissAllele = NULL,
  splitPhasedReads = FALSE,
  compareSingleCells = FALSE,
  exportVCF = NULL,
  bsGenome = NULL,
  ref.fasta = NULL,
  assume.biallelic = FALSE
)
}
\arguments{
\item{inputfolder}{Path to the bam files to process}

\item{outputfolder}{Output directory. If non-existent it will be created.}

\item{positions}{Filename with listed position of SNVs for given chromosome (format: chrName SNVpos).}

\item{WCregions}{Filename of all WC region for a given chromosome (format: chrName:Start:End:FileName).}

\item{chromosome}{If only a subset of the chromosomes should be processed, specify them here.}

\item{pairedEndReads}{Set to \code{TRUE} if you have paired-end reads in your file.}

\item{min.mapq}{Minimum mapping quality when importing from BAM files.}

\item{min.baseq}{Minimum base quality to consider a base for phasing.}

\item{num.iterations}{Number of iteration to sort watson and crick matrices.}

\item{translateBases}{translates integer coded bases (1,2,3,4) into letters (A,C,G,T)}

\item{concordance}{Level of agreement between single cell and consensus haplotypes}

\item{fillMissAllele}{A patch to a single BAM or VCF file for a given sample to be used to fill missing alleles, uncovered in Strand-seq data.}

\item{splitPhasedReads}{Set to \code{TRUE} if you want to split reads per haplotype.}

\item{compareSingleCells}{Set to \code{TRUE} if you want to compare haplotypes at the single-cell level.}

\item{exportVCF}{Ideally a sample ID that if defined invokes export of phased haplotypes in a separate VCF file.}

\item{bsGenome}{A \code{BSgenome} object which contains reference genome used to infer reference alleles.}

\item{ref.fasta}{A user defined reference FASTA file to extract reference allele for all SNV positions.}

\item{assume.biallelic}{If set to \code{TRUE} parameter 'snv.positions' is expected to contain biallelic loci (0/1, 1/0) and thus
gaps in haplotypes will be filled accordingly.}
}
\description{
This function will move through .bam files in a folder and perform several steps (see Details).
}
\details{
1. extract variable position in WC regions
2. Fill two matrices separately for SNVs found in Watson and Crick reads
3. Sort matrices in order each column in each matrix has lowest amount of conflicting bases
4. Exclude rows/cells which cannot be reliably assigned to only one matrix consensus 
5. For successfully phased rows/cell export W and C reads as a separate haplotype specifiv GRanges object
}
\author{
David Porubsky
}
