% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctInversionPhasing.R
\name{phaseHETinversion}
\alias{phaseHETinversion}
\title{Phase a heterozygous inversion using Strand-seq data}
\usage{
phaseHETinversion(
  input.bams = NULL,
  snv.positions = NULL,
  phase.gr = NULL,
  lookup.bp = 1e+06,
  lookup.blacklist = NULL,
  pairedEndReads = TRUE,
  min.mapq = 10,
  bsGenome = NULL,
  ref.fasta = NULL,
  assume.biallelic = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{input.bams}{A path to a VCF file to be loaded.}

\item{snv.positions}{A path to a VCF files containing SNV positions to phase.}

\item{phase.gr}{A \code{\link{GRanges}} object of region to be phased. This region should point to a heterozygous inversion site.}

\item{lookup.bp}{A number of nucleotides, downstream and upstream, from the heterozygous inversion site ('phase.gr') to be genotyped.}

\item{lookup.blacklist}{A \code{\link{GRanges}} object or a path to a BED file containing a set of ranges to be excluded 
when extending 'phase.gr' by 'lookup.bp'. The total size of 'lookup.bp' is kept after filtering.}

\item{pairedEndReads}{Set to \code{TRUE} if you have paired-end reads in your file.}

\item{min.mapq}{Minimum mapping quality when importing from BAM files.}

\item{bsGenome}{A \code{BSgenome} object which contains reference genome used to infer reference alleles.}

\item{ref.fasta}{A user defined reference FASTA file to extract reference allele for all SNV positions.}

\item{assume.biallelic}{If set to \code{TRUE} parameter 'snv.positions' is expected to contain biallelic loci (0/1, 1/0) and thus
gaps in haplotypes will be filled accordingly.}

\item{verbose}{Is set to \code{TRUE} function will provide a more detailed messaging of ongoing analysis steps.}
}
\value{
A \code{data.frame} with phased alleles per inverted and reference haplotype.
}
\description{
This function takes as an input region deemed to be a heterozygous inversion and attempts to phase SNVs inside this region using
haplotype informative Strand-seq reads from multiple single cells.
}
\author{
David Porubsky
}
