% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotHapDensity}
\alias{plotHapDensity}
\title{This function will take phased data for each chromosome and plot density of phased SNVs}
\usage{
plotHapDensity(datapath, perChromosome = FALSE, file = NULL,
  HighQual = FALSE)
}
\arguments{
\item{datapath}{containing split watson and crick reads per haplotype}

\item{perChromosome}{if TRUE function will plot density for every chromosome, otherwise one summary density plot is exported}

\item{file}{name of the file to save produced plots (add appropriate file extension)}
}
\description{
This function will take phased data for each chromosome and plot density of phased SNVs
}
\author{
David Porubsky
}

