% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotHaploSegs}
\alias{plotHaploSegs}
\title{Plot inherited haplotype segments as ideogram.}
\usage{
plotHaploSegs(hapSegm.gr = NULL, layout = "vertical")
}
\arguments{
\item{hapSegm.gr}{A \code{\link{GRanges-class}} object with regions of each inherited haplotype block}

\item{layout}{Set to 'vertical' or 'horizonal' based on desired layout of an resultant ideogram.}
}
\value{
A \code{ggplot} object.
}
\description{
This function takes as na input a \code{\link{GRanges-class}} object with genomic positions of each
inherited haplotype block and project them as whole-genome ideogram.
}
\author{
David Porubsky
}
