% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotPhasedReads}
\alias{plotPhasedReads}
\title{This function will take phased reads for each chromosome and plot them}
\usage{
plotPhasedReads(datapath, plotChromosomes = NULL, file = NULL)
}
\arguments{
\item{plotChromosomes}{}

\item{file}{name of the file to save produced plots}

\item{data.object}{containing split watson and crick reads per haplotype}
}
\description{
This function will take phased reads for each chromosome and plot them
}
\author{
David Porubsky
}

