% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotSingleCellHaps}
\alias{plotSingleCellHaps}
\title{This function will take phased data for each chromosome and plot density of phased SNVs}
\usage{
plotSingleCellHaps(data, file = NULL)
}
\arguments{
\item{data}{A \code{list} output from the function \code{\link{compareSingleCellHaps}}}

\item{file}{name of the file to save produced plots (add appropriate file extension)}
}
\description{
This function will take phased data for each chromosome and plot density of phased SNVs
}
\author{
David Porubsky
}
