% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runStrandPhaseR.R
\name{runStrandPhaseR}
\alias{runStrandPhaseR}
\title{Wrapper function for StrandPhaseR}
\usage{
runStrandPhaseR(bamfilespath, dataDirectory = "./StrandPhaseR_analysis",
  positions = NULL, WCregions = NULL, chromosomes = NULL,
  pairedEndReads = TRUE, min.mapq = 10, min.baseq = 30,
  num.iterations = 2, translateBases = TRUE, score2qual = FALSE,
  fillGaps = NULL, splitPhasedReads = FALSE, callBreaks = FALSE)
}
\arguments{
\item{bamfilespath}{Path to the bam files to process}

\item{dataDirectory}{Output directory. If non-existent it will be created.}

\item{positions}{Filename with listed position of SNVs for given chromosome (format: chrName SNVpos).}

\item{WCregions}{Filename of all WC region for a given chromosome (format: chrName:Start:End:FileName).}

\item{chromosomes}{If only a subset of the chromosomes should be processed, specify them here.}

\item{pairedEndReads}{Set to \code{TRUE} if you have paired-end reads in your file.}

\item{min.mapq}{Minimum mapping quality when importing from BAM files.}

\item{min.baseq}{Minimum base quality to consider a base for phasing.}

\item{num.iterations}{Number of iteration to sort watson and crick matrices.}

\item{splitPhasedReads}{Set to \code{TRUE} if you want to split reads per haplotype.}
}
\description{
This function will move through .bam files in a folder and perform several steps (see Details).
}
\details{
1. extract variable position in WC regions
2. Fill two matrices separately for SNVs found in Watson and Crick reads
3. Sort matrices in order each column in each matrix has lowest amount of conflicting bases
4. Exclude rows/cells which cannot be reliably assigned to only one matrix consensus
5. For successfully phased rows/cell export W and C reads as a separate haplotype specifiv GRanges object
}
\author{
David Porubsky
}

