% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctInversionPhasing.R
\name{sliceRanges}
\alias{sliceRanges}
\title{Slice a certain length from a set of genomic ranges.}
\usage{
sliceRanges(gr, slice.size = 1e+06, from = "start")
}
\arguments{
\item{gr}{A \code{\link{GRanges}} object containing a set of genomic ranges.}

\item{slice.size}{A total length of sliced ranges to be reported given the starting range defined by 'from'.}

\item{from}{Set to 'start' or 'end' if to start counting bases to slice from the first or the last range.}
}
\value{
A \code{\link{GRanges}} object.
}
\description{
This function takes as an input a set of ranges stored in \code{\link{GRanges}} object and export slice of certain length
starting either from the first or last range.
}
\author{
David Porubsky
}
