% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortMatrices.R
\name{sortMatrices}
\alias{sortMatrices}
\title{This funcion will load initial watson and crick matrices and will sort them according to the phase information
of each single cell}
\usage{
sortMatrices(data.object, num.iterations = 2)
}
\arguments{
\item{data.object}{containing watson and crick haplotypes of each single cell}

\item{num.iterations}{number of iteration to sort watson and crick matrices}
}
\description{
This funcion will load initial watson and crick matrices and will sort them according to the phase information
of each single cell
}
\author{
David Porubsky
}
