% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StrandPhase.R
\name{strandPhaseR}
\alias{strandPhaseR}
\title{Wrapper function}
\usage{
strandPhaseR(
  inputfolder,
  outputfolder = "./StrandPhaseR_analysis",
  configfile = NULL,
  numCPU = 1,
  positions = NULL,
  WCregions = NULL,
  chromosomes = NULL,
  pairedEndReads = TRUE,
  min.mapq = 10,
  min.baseq = 20,
  num.iterations = 2,
  translateBases = TRUE,
  concordance = 0.9,
  fillMissAllele = NULL,
  splitPhasedReads = FALSE,
  compareSingleCells = FALSE,
  exportVCF = NULL,
  bsGenome = NULL
)
}
\arguments{
\item{inputfolder}{Path to the bam files to process}

\item{outputfolder}{Output directory. If non-existent it will be created.}

\item{numCPU}{The numbers of CPUs that are used. Should not be more than available on your machine.}

\item{positions}{Filename with listed position of SNVs for given chromosome (format: chrName SNVpos).}

\item{WCregions}{Filename of all WC region for a given chromosome (format: chrName:Start:End:FileName).}

\item{pairedEndReads}{Set to \code{TRUE} if you have paired-end reads in your file.}

\item{min.mapq}{Minimum mapping quality when importing from BAM files.}

\item{min.baseq}{Minimum base quality to consider a base for phasing.}

\item{num.iterations}{Number of iteration to sort watson and crick matrices.}

\item{translateBases}{translates integer coded bases (1,2,3,4) into letters (A,C,G,T)}

\item{concordance}{Level of agreement between single cell and consensus haplotypes}

\item{splitPhasedReads}{Set to \code{TRUE} if you want to split reads per haplotype.}

\item{compareSingleCells}{Set to \code{TRUE} if you want to compare haplotypes at the single-cell level.}

\item{exportVCF}{...}

\item{bsGenome}{A \code{BSgenome} object which contains reference genome used to infer reference alleles.}
}
\description{
This function will move through .bam files in a folder and perform several steps (see Details).
}
\author{
David Porubsky
}
