% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2ranges.R
\name{vcf2ranges}
\alias{vcf2ranges}
\title{Read VCF file into a \code{\link{GRanges}} object}
\usage{
vcf2ranges(vcfFile = NULL, genotypeField = 1, chromosomes = NULL)
}
\arguments{
\item{vcfFile}{VCF file containing phased alleles}

\item{genotypeField}{In case of multiple samples phased in a single file (Default: 1)}
}
\description{
Read VCF file into a \code{\link{GRanges}} object
}
\author{
David Porubsky
}

