% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembleHaps.R
\name{assembleHaps}
\alias{assembleHaps}
\title{This will take sorted matrices and will calculate concordance of each single cell to the consensus haplotypes in order to assemble highly 
accurate haplotypes}
\usage{
assembleHaps(data.object, translateBases = FALSE)
}
\arguments{
\item{data.object}{containing sorted watson and crick haplotypes of each single cell}
}
\description{
This will take sorted matrices and will calculate concordance of each single cell to the consensus haplotypes in order to assemble highly 
accurate haplotypes
}
\author{
David Porubsky
}

