% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{exportBedGraph}
\alias{exportBedGraph}
\title{Generates a bedgraph from GRanges object in order to upload on to UCSC Genome browser}
\usage{
exportBedGraph(index, outputfolder, fragments = NULL, col = "200,100,10")
}
\arguments{
\item{index}{is used to index the bedfile(s)}

\item{outputfolder}{Location to store created bedfile(s)}

\item{fragments}{A \code{\link{GRanges}} object with strand and mapq metadata, such as that generated by \code{\link{bam2GRanges}}}
}
\description{
Generates a bedgraph from GRanges object in order to upload on to UCSC Genome browser
}
\author{
David Porubsky, Ashley Sanders
}
