% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{exportVCF}
\alias{exportVCF}
\title{Generates a VCF file from phased haplotypes}
\usage{
exportVCF(index = NULL, outputfolder, phasedHap = NULL, bsGenome,
  chromosome = chr)
}
\arguments{
\item{index}{Unique identifier used to index analyzed individual/sample}

\item{outputfolder}{Location to store created VCF file(s)}

\item{phasedHap}{Data object containing phased haplotypes}

\item{bsGenome}{A \code{BSgenome} object which contains reference genome usedto infer reference bases}

\item{chromosome}{Name of the chromosome for which we want to export vcf file}
}
\description{
Generates a VCF file from phased haplotypes
}
\author{
David Porubsky, Ashley Sanders
}
