% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{exportVCF}
\alias{exportVCF}
\title{Generates a VCF file from phased haplotypes}
\usage{
exportVCF(index = NULL, outputfolder = NULL, phasedHap = NULL,
  positions = NULL, bsGenome = NULL, chromosome = NULL)
}
\arguments{
\item{index}{Unique identifier used to index analyzed individual/sample}

\item{outputfolder}{Location to store created VCF file(s)}

\item{phasedHap}{Data object containing phased haplotypes}

\item{positions}{Set of heterozygous SNV positions used for phasing obtained from an input VCF files.}

\item{bsGenome}{A \code{BSgenome} object which contains reference genome used to infer reference alleles.}

\item{chromosome}{Name of the chromosome for which we want to export vcf file}
}
\description{
Generates a VCF file from phased haplotypes
}
\author{
David Porubsky
}
