% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillGaps.R
\name{fillGaps}
\alias{fillGaps}
\title{This function will take both assembled haplotypes and will try to fill gaps at positions where only one allele is phased
Such position have to be heterozygous so the alternative allele at this position can be reliably distinguished}
\usage{
fillGaps(data.object, merged.bam, min.mapq = 10, min.baseq = 30,
  translateBases = FALSE, chromosome = NULL, chunkSize = 1e+07,
  filterAltAlign = TRUE)
}
\description{
This function will take both assembled haplotypes and will try to fill gaps at positions where only one allele is phased
Such position have to be heterozygous so the alternative allele at this position can be reliably distinguished
}
\author{
David Porubsky
}

