% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/splitReads.R
\name{splitReads}
\alias{splitReads}
\title{This will take phased info for each haplotype and will split directional reads of each single cell into haplotype
specific reads}
\usage{
splitReads(data.object = NULL, bamfilespath = bamfilespath,
  pairedEndReads = FALSE, min.mapq = 10)
}
\arguments{
\item{data.object}{containing sorted and filtered watson and crick haplotypes of each single cell along with phase info}

\item{gr.list}{}
}
\description{
This will take phased info for each haplotype and will split directional reads of each single cell into haplotype
specific reads
}
\author{
David Porubsky
}

