% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitReads.R
\name{splitReads}
\alias{splitReads}
\title{This will take phased info for each haplotype and will split directional reads of each single cell into haplotype
specific reads}
\usage{
splitReads(
  data.object = NULL,
  inputfolder = inputfolder,
  pairedEndReads = FALSE,
  min.mapq = 10,
  filterAltAlign = TRUE
)
}
\arguments{
\item{data.object}{containing sorted and filtered watson and crick haplotypes of each single cell along with phase info}

\item{inputfolder}{Path to the bam files to process}

\item{pairedEndReads}{Set to \code{TRUE} if you have paired-end reads in your file.}

\item{min.mapq}{Minimum mapping quality when importing from BAM files.}
}
\description{
This will take phased info for each haplotype and will split directional reads of each single cell into haplotype
specific reads
}
\author{
David Porubsky
}
