% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/StrandPhase.R
\name{strandPhaseR}
\alias{strandPhaseR}
\title{Wrapper function}
\usage{
strandPhaseR(inputfolder, outputfolder = "./StrandPhaseR_analysis",
  configfile = NULL, numCPU = 1, positions = NULL, WCregions = NULL,
  chromosomes = NULL, pairedEndReads = TRUE, min.mapq = 10,
  min.baseq = 30, num.iterations = 2, translateBases = TRUE,
  fillMissAllele = NULL, splitPhasedReads = FALSE, callBreaks = FALSE,
  exportVCF = NULL, bsGenome = NULL)
}
\arguments{
\item{inputfolder}{Path to the bam files to process}

\item{outputfolder}{Output directory. If non-existent it will be created.}

\item{numCPU}{The numbers of CPUs that are used. Should not be more than available on your machine.}

\item{positions}{Filename with listed position of SNVs for given chromosome (format: chrName SNVpos).}

\item{WCregions}{Filename of all WC region for a given chromosome (format: chrName:Start:End:FileName).}

\item{chromosomes}{If only a subset of the chromosomes should be processed, specify them here.}

\item{pairedEndReads}{Set to \code{TRUE} if you have paired-end reads in your file.}

\item{min.mapq}{Minimum mapping quality when importing from BAM files.}

\item{min.baseq}{Minimum base quality to consider a base for phasing.}

\item{num.iterations}{Number of iteration to sort watson and crick matrices.}

\item{translateBases}{}

\item{splitPhasedReads}{Set to \code{TRUE} if you want to split reads per haplotype.}

\item{callBreaks}{}

\item{exportVCF}{}

\item{bsGenome}{}
}
\description{
This function will move through .bam files in a folder and perform several steps (see Details).
}
\author{
David Porubsky
}

